Alternate isearch behavior
The original behavior during isearch was to immediately show all folded
regions as soon as isearch was started. I found this to be disorienting.
This commit adds a customization option to change this behavior so that
it can follow a more standard isearch behavior of temporarily unhiding a
region as the isearch enters that region, then re-hiding it when the
search leaves the region.

(I am observing some weird behavior when I toggle isearch to only search
visible text with M-s i: If I'm searching forward, the cursor makes a
single stop near the ellipsis of a folded region if there are any
matches in that region, though without unfolding the region. But if
searching backward, this doesn't happen: the folded regions are skipped
over, as expected. I'm thinking this may be a bug somewhere else in
Emacs, outside of yafolding. However, even with this odd behavior, I
still prefer this new behavior over the original yafolding behavior;
and in any case, I typically search with search-invisible set to the
standard value of 'open.)